requirements.txt
streamlit
langchain
langchain-community
langchain-groq
faiss-cpu
python-docx
python-dotenv
huggingface-hub
sentence-transformers
requests
langchain-text-splitters
langchain-huggingface